traductor proyecto solitario/requirements.txt
tk
SpeechRecognition
pyaudio
google-generativeai
python-dotenv
gTTS